+++ pkg/secrets/get_secrets.go
@@ -8,6 +8,7 @@ import (
 	"github.com/boringtools/git-alerts/internal/ui"
 	"github.com/boringtools/git-alerts/pkg/common"
 	"github.com/boringtools/git-alerts/pkg/models"
+	"github.com/go-git/go-git/v5/plumbing/transport"
 )
 
 var (
@@ -23,10 +24,20 @@ func RunSecretsScan() {
 		for key, value := range repo {
 			if !value.Fork {
 				cloneDirectory := filepath.Join(common.CloneDirectoryPath, strconv.Itoa(key))
-				CloneRepo(value.CloneURL, cloneDirectory)
-				ui.PrintMsg("Scanning repository : %s", value.CloneURL)
-				RunGitleaks(cloneDirectory, true)
-				common.RemoveDirectory(cloneDirectory)
+				errClone := CloneRepo(value.CloneURL, cloneDirectory)
+
+				if errClone != nil {
+					if errClone == transport.ErrEmptyRemoteRepository {
+						ui.PrintWarning("Skipping empty repository : %s", value.CloneURL)
+					} else {
+						ui.PrintError("Error while cloning repository : %s", errClone)
+					}
+				} else {
+					ui.PrintMsg("Scanning repository : %s", value.CloneURL)
+					RunGitleaks(cloneDirectory, true)
+					common.RemoveDirectory(cloneDirectory)
+				}
+
 			} else {
 				ui.PrintWarning("Skipping forked repository : %s", value.CloneURL)
 			}
